Move the non-analytic correction engine into kaldo/controllers/nac.py
The dipole-dipole terms, supercell mapping, short-range reconstruction and
dynamical-matrix assembly lived as module-level helpers inside
secondorder.py, with the velocity constants and unit converters in
harmonic_with_q.py. They are one self-contained engine; give them their own
controller module. SecondOrder keeps thin cached accessors and
HarmonicWithQ keeps thin dispatch; the moved names are temporarily
re-exported from secondorder for existing imports and get renamed in the
following commit.
